+++ backend/baton/src/main/java/touch/baton/domain/runnerpost/query/service/RunnerPostQueryService.java
@@ -3,6 +3,7 @@ package touch.baton.domain.runnerpost.query.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import touch.baton.domain.common.request.PageParams;
 import touch.baton.domain.common.response.PageResponse;
 import touch.baton.domain.member.command.Runner;
 import touch.baton.domain.member.command.SupporterRunnerPost;
@@ -15,7 +16,6 @@ import touch.baton.domain.runnerpost.command.vo.ReviewStatus;
 import touch.baton.domain.runnerpost.query.controller.response.RunnerPostResponse;
 import touch.baton.domain.runnerpost.query.repository.RunnerPostPageRepository;
 import touch.baton.domain.runnerpost.query.repository.RunnerPostQueryRepository;
-import touch.baton.domain.runnerpost.query.service.dto.PageParams;
 import touch.baton.domain.tag.command.RunnerPostTag;
 import touch.baton.domain.tag.command.vo.TagReducedName;
 import touch.baton.domain.tag.query.repository.RunnerPostTagQueryRepository;
