+++ backend/baton/src/test/java/touch/baton/document/runnerpost/read/RunnerPostReadWithLoginedRunnerApiTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import touch.baton.config.RestdocsConfig;
+import touch.baton.domain.common.request.PageParams;
 import touch.baton.domain.common.response.PageResponse;
 import touch.baton.domain.member.command.Member;
 import touch.baton.domain.member.command.Runner;
@@ -15,7 +16,6 @@ import touch.baton.domain.runnerpost.command.vo.ReviewStatus;
 import touch.baton.domain.runnerpost.query.controller.RunnerPostQueryController;
 import touch.baton.domain.runnerpost.query.controller.response.RunnerPostResponse;
 import touch.baton.domain.runnerpost.query.service.RunnerPostQueryService;
-import touch.baton.domain.runnerpost.query.service.dto.PageParams;
 import touch.baton.domain.tag.command.Tag;
 import touch.baton.fixture.domain.MemberFixture;
 import touch.baton.fixture.domain.RunnerFixture;
